+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Capability to generate documentation for crossplane XRDs
+- Capability to generate documentation for kratix promises
+
 ## [0.11.1] - 2024-02-21
 
 ### Changed